DEV: add tools.py to separates functions depends on xpdacq.glbl
